+++ packages/core/src/date-range-picker/date-range-picker-root.vue
@@ -1,14 +1,12 @@
 <script lang="ts">
 import type { DateValue } from '@internationalized/date';
-
 import type { Ref } from 'vue';
 import type { ADateRangeFieldRoot, ADateRangeFieldRootProps, APopoverRootEmits, APopoverRootProps, ARangeCalendarRootProps } from '..';
 import type { DateMatcher, WeekDayFormat } from '../date';
 import type { DateRange, DateStep, Granularity, HourCycle } from '../shared/date';
-
 import type { Direction } from '../shared/types';
 import { APopoverRoot } from '..';
-import { createContext, useDirection } from '../shared';
+import { createContext, useDirection, useLocale } from '../shared';
 import { getDefaultDate } from '../shared/date';
 
 type DateRangePickerRootContext = {
@@ -54,7 +52,7 @@ export type ADateRangePickerRootProps = Omit<ADateRangeFieldRootProps, 'as' | 'a
   closeOnSelect?: boolean;
 };
 
-export type ADateRangePickerRootEmits = {
+export type ADateRangePickerRootEmits = APopoverRootEmits & {
   /** Event handler called whenever the model value changes */
   'update:modelValue': [date: DateRange];
   /** Event handler called whenever the placeholder value changes */
@@ -93,7 +91,6 @@ const props = withDefaults(
     disabled: false,
     readonly: false,
     placeholder: undefined,
-    locale: 'en',
     isDateDisabled: undefined,
     isDateUnavailable: undefined,
     isDateHighlightable: undefined,
@@ -103,10 +100,10 @@ const props = withDefaults(
   },
 );
 
-const emits = defineEmits<ADateRangePickerRootEmits & APopoverRootEmits>();
+const emits = defineEmits<ADateRangePickerRootEmits>();
 
 const {
-  locale,
+  locale: propLocale,
   disabled,
   readonly,
   pagedNavigation,
@@ -137,6 +134,7 @@ const {
 } = toRefs(props);
 
 const dir = useDirection(propsDir);
+const locale = useLocale(propLocale);
 
 const modelValue = useVModel(props, 'modelValue', emits, {
   defaultValue: props.defaultValue ?? { start: undefined, end: undefined },
@@ -147,7 +145,7 @@ const defaultDate = getDefaultDate({
   defaultPlaceholder: props.placeholder,
   granularity: props.granularity,
   defaultValue: modelValue.value?.start,
-  locale: props.locale,
+  locale: locale.value,
 });
 
 const placeholder = useVModel(props, 'placeholder', emits, {