+++ packages/core/src/date/comparators.date.ts
@@ -2,7 +2,6 @@ import type { DateValue } from '@internationalized/date';
 import type { DateMatcher } from './types.date';
 import {
   CalendarDateTime,
-
   getDayOfWeek,
   getLocalTimeZone,
   parseDate,
@@ -175,7 +174,11 @@ export function getLastFirstDayOfWeek<T extends DateValue = DateValue>(
   { date, firstDayOfWeek, locale }:
   DayOfWeekParams<T>,
 ): T {
-  const day = getDayOfWeek(date, locale);
+  /**
+   * "firstDayOfWeek" is fixed to 0(Sunday) to avoid confusion regarding locales.
+   * This also aligns with other date libraries, e.g., date-fns.
+   */
+  const day = getDayOfWeek(date, locale, 'sun');
 
   if (firstDayOfWeek > day) {
     return date.subtract({ days: day + 7 - firstDayOfWeek }) as T;
@@ -192,7 +195,12 @@ export function getNextLastDayOfWeek<T extends DateValue = DateValue>(
   { date, firstDayOfWeek, locale }:
   DayOfWeekParams<T>,
 ): T {
-  const day = getDayOfWeek(date, locale);
+  /**
+   * "firstDayOfWeek" is fixed to 0(Sunday) to avoid confusion regarding locales.
+   * This also aligns with other date libraries, e.g., date-fns.
+   */
+  const day = getDayOfWeek(date, locale, 'sun');
+
   const lastDayOfWeek = firstDayOfWeek === 0 ? 6 : firstDayOfWeek - 1;
 
   if (day === lastDayOfWeek) {