+++ packages/core/src/time-field/time-field-input.vue
@@ -23,6 +23,7 @@ const lastKeyZero = ref(false);
 const {
   handleSegmentClick,
   handleSegmentKeydown,
+  handleSegmentFocusOut,
   attributes,
 } = useDateField({
   hasLeftFocus,
@@ -30,6 +31,7 @@ const {
   placeholder: rootContext.placeholder,
   hourCycle: rootContext.hourCycle,
   step: rootContext.step,
+  stepSnapping: rootContext.stepSnapping,
   segmentValues: rootContext.segmentValues,
   formatter: rootContext.formatter,
   part: props.part,
@@ -59,9 +61,12 @@ const isInvalid = computed(() => rootContext.isInvalid.value);
     v-on="part !== 'literal' ? {
       mousedown: handleSegmentClick,
       keydown: handleSegmentKeydown,
-      focusout: () => { hasLeftFocus = true },
+      focusout: () => {
+        hasLeftFocus = true;
+        handleSegmentFocusOut();
+      },
       focusin: (event: FocusEvent) => {
-        rootContext.setFocusedElement(event.target as HTMLElement)
+        rootContext.setFocusedElement(event.target as HTMLElement);
       },
     } : {}"
   >