+++ packages/pohon/CHANGELOG.md
@@ -13,7 +13,7 @@
   - feat(PopperContent): add hideShiftedArrow prop to control arrow visibility
   - fix(DateRangePicker): add exports for missing component types and include TimeValue type in exports
   - fix(PinInput): handle more edge cases for the placeholder
-  - fix: dayPeriod returning incorrect value due to missmatched format
+  - fix: dayPeriod returning incorrect value due to mismatched format
   - fix(Editable): prevent submission during IME composition
   - fix(NavigationMenu): allow aria-label override by fixing CollectionSlot attrs forwarding
   - fix(Avatar): Properly set referrerpolicy and crossorigin